src/frontend/requirements.txt
PySide6>=6.5.0
psutil>=5.9.0
pynvml>=11.5.0  # For GPU monitoring
mistralai>=0.0.7  # For Mistral LLM integration
python-dotenv>=1.0.0  # For environment variable management
black>=23.3.0  # For code formatting
ruff>=0.0.270  # For linting
pytest>=7.3.1  # For testing 